requirements.txt
streamlit
requests
python-dotenv
ollama


